# Conflicts:
#	modules/javafx.graphics/src/main/java/javafx/scene/Node.java

+++ modules/javafx.base/src/main/java/javafx/util/converter/FloatStringConverter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2010, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -63,6 +63,6 @@ public class FloatStringConverter extends StringConverter<Float> {
             return "";
         }
 
-        return Float.toString(((Float)value).floatValue());
+        return Float.toString(value.floatValue());
     }
 }
